+++ changelog.md
@@ -1,3 +1,1 @@
-- Fix fluid ingredient codec in debug mode
-- Fix data module entry points not existing on fabric
-- Update energy lib
+- Fix nullability issue with Fabric energy storage